CP-54351 move cifutils from /opt
